requirements.txt
streamlit
pandas
python-dotenv
langchain
langchain_openai
langchain_experimental
tabulate
openai